Evaluators

Description

Argos.AbstractNLPEvaluator — Type

AbstractNLPEvaluator

AbstractNLPEvaluator implements the bridge between the problem formulation (see ExaPF.AbstractFormulation) and the optimization solver. Once the problem formulation bridged, the evaluator allows to evaluate:

the objective;

the gradient of the objective;

the constraints;

the Jacobian of the constraints;

the Jacobian-vector and transpose-Jacobian vector products of the constraints;

the Hessian of the objective;

the Hessian of the Lagrangian.

API Reference

Optimization

Argos.optimize! — Function

optimize!(optimizer, nlp::AbstractNLPEvaluator, x0)

Use optimization routine implemented in optimizer to optimize the optimal power flow problem specified in the evaluator nlp. Initial point is specified by x0.

Return the solution as a named tuple, with fields

status::MOI.TerminationStatus: Solver's termination status, as specified by MOI

minimum::Float64: final objective

minimizer::AbstractVector: final solution vector, with same ordering as the Variables specified in nlp.

optimize!(optimizer, nlp::AbstractNLPEvaluator)

Wrap previous optimize! function and pass as initial guess x0 the initial value specified when calling initial(nlp).

Examples

nlp = ExaPF.ReducedSpaceEvaluator(datafile)
optimizer = Ipopt.Optimizer()
solution = ExaPF.optimize!(optimizer, nlp)

Notes

By default, the optimization routine solves a minimization problem.

Attributes

Argos.Variables — Type

Variables <: AbstractNLPAttribute end

Attribute corresponding to the optimization variables attached to a given AbstractNLPEvaluator.

Argos.Constraints — Type

Constraints <: AbstractNLPAttribute end

Attribute corresponding to the constraints attached to a given AbstractNLPEvaluator.

Argos.n_variables — Function

n_variables(nlp::AbstractNLPEvaluator)

Get the number of variables in the problem.

Argos.n_constraints — Function

n_constraints(nlp::AbstractNLPEvaluator)

Get the number of constraints in the problem.

Argos.constraints_type — Function

constraints_type(nlp::AbstractNLPEvaluator)

Return the type of the non-linear constraints of the evaluator nlp, as a Symbol. Result could be :inequality if problem has only inequality constraints, :equality if problem has only equality constraints, or :mixed if problem has both types of constraints.

Utilities

Argos.reset! — Function

reset!(nlp::AbstractNLPEvaluator)

Reset evaluator nlp to default configuration.

Missing docstring.

Missing docstring for primal_infeasibility. Check Documenter's build log for details.

Missing docstring.

Missing docstring for primal_infeasibility!. Check Documenter's build log for details.

Callbacks

Objective

Argos.gradient! — Function

gradient!(nlp::AbstractNLPEvaluator, g, u)

Evaluate the gradient of the objective, at given variable u. Store the result inplace in the vector g.

Note

The vector g should have the same dimension as u.

Argos.hessprod! — Function

hessprod!(nlp::AbstractNLPEvaluator, hessvec, u, v)

Evaluate the Hessian-vector product ∇²f(u) * v of the objective evaluated at variable u. Store the result inplace, in the vector hessvec.

Note

The vector hessprod should have the same length as u.

Argos.hessian! — Function

hessian!(nlp::AbstractNLPEvaluator, H, u)

Evaluate the Hessian ∇²f(u) of the objective function f(u). Store the result inplace, in the n x n dense matrix H.

Constraints

Argos.constraint! — Function

constraint!(nlp::AbstractNLPEvaluator, cons, u)

Evaluate the constraints of the problem at given variable u. Store the result inplace, in the vector cons.

Note

The vector cons should have the same dimension as the result returned by n_constraints(nlp).

Missing docstring.

Missing docstring for jacobian_structure!. Check Documenter's build log for details.

Argos.jacobian! — Function

jacobian!(nlp::AbstractNLPEvaluator, jac::AbstractMatrix, u)

Evaluate the Jacobian of the constraints, at variable u. Store the result inplace, in the m x n dense matrix jac.

Argos.jprod! — Function

jprod!(nlp::AbstractNLPEvaluator, jv, u, v)

Evaluate the Jacobian-vector product $J v$ of the constraints. The vector jv is modified inplace.

Let (n, m) = n_variables(nlp), n_constraints(nlp).

u is a vector with dimension n

v is a vector with dimension n

jv is a vector with dimension m

Argos.jtprod! — Function

jtprod!(nlp::AbstractNLPEvaluator, jv, u, v)

Evaluate the transpose Jacobian-vector product $J^{T} v$ of the constraints. The vector jv is modified inplace.

Let (n, m) = n_variables(nlp), n_constraints(nlp).

u is a vector with dimension n

v is a vector with dimension m

jv is a vector with dimension n

Argos.ojtprod! — Function

ojtprod!(nlp::AbstractNLPEvaluator, jv, u, σ, v)

Evaluate the transpose Jacobian-vector product J' * [σ ; v], with J the Jacobian of the vector [f(x); h(x)]. f(x) is the current objective and h(x) constraints. The vector jv is modified inplace.

Let (n, m) = n_variables(nlp), n_constraints(nlp).

jv is a vector with dimension n

u is a vector with dimension n

σ is a scalar

v is a vector with dimension m

Second-order

Argos.hessian_lagrangian_penalty_prod! — Function

hessian_lagrangian_penalty_prod!(nlp::AbstractNLPEvaluator, hessvec, u, y, σ, d, v)

Evaluate the Hessian-vector product of the Augmented Lagrangian function $L(u, y) = f(u) + \sum_i y_i c_i(u) + \frac{1}{2} d_i c_i(u)^2$ with a vector v:

\[∇²L(u, y) ⋅ v = σ ∇²f(u) ⋅ v + \sum_i (y_i + d_i) ∇²c_i(u) ⋅ v + \sum_i d_i ∇c_i(u)^T ∇c_i(u)\]

Store the result inplace, in the vector hessvec.

Arguments

hessvec is a AbstractVector with dimension n, which is modified inplace.

u is a AbstractVector with dimension n, storing the current variable.

y is a AbstractVector with dimension n, storing the current constraints' multipliers

σ is a scalar

v is a vector with dimension n.

d is a vector with dimension m.

ReducedSpaceEvaluator

When working in the reduced space, we could use the corresponding ReducedSpaceEvaluator:

Argos.ReducedSpaceEvaluator — Type

ReducedSpaceEvaluator{T, VI, VT, MT, Jacx, Jacu, JacCons, Hess} <: AbstractNLPEvaluator

Reduced-space evaluator projecting the optimization problem into the powerflow manifold defined by the nonlinear equation $g(x, u) = 0$. The state $x$ is defined implicitly, as a function of the control $u$. Hence, the powerflow equation is implicitly satisfied when we are using this evaluator.

Once a new point u is passed to the evaluator, the user needs to call the method update! to find the corresponding state $x(u)$ satisfying the balance equation $g(x(u), u) = 0$.

Taking as input an ExaPF.PolarForm structure, the reduced evaluator builds the bounds corresponding to the control u, The reduced evaluator could be instantiated on the host memory, or on a specific device (currently, only CUDA is supported).

Examples

julia> datafile = "case9.m"  # specify a path to a MATPOWER instance
julia> nlp = ReducedSpaceEvaluator(datafile)
A ReducedSpaceEvaluator object
    * device: KernelAbstractions.CPU()
    * #vars: 5
    * #cons: 10
    * constraints:
        - voltage_magnitude_constraints
        - active_power_constraints
        - reactive_power_constraints
    * linear solver: ExaPF.LinearSolvers.DirectSolver()

If a GPU is available, we could instantiate nlp as

julia> nlp_gpu = ReducedSpaceEvaluator(datafile; device=CUDADevice())
A ReducedSpaceEvaluator object
    * device: KernelAbstractions.CUDADevice()
    * #vars: 5
    * #cons: 10
    * constraints:
        - voltage_magnitude_constraints
        - active_power_constraints
        - reactive_power_constraints
    * linear solver: ExaPF.LinearSolvers.DirectSolver()

Note

Mathematically, we set apart the state $x$ from the control $u$, and use a third variable $y$ –- the by-product –- to denote the remaining values of the network. In the implementation of ReducedSpaceEvaluator, we only deal with a control u and an attribute buffer, storing all the physical values needed to describe the network. The attribute buffer stores the values of the control u, the state x and the by-product y. Each time we are calling the method update!, the values of the control are copied into the buffer.

SlackEvaluator

Argos.SlackEvaluator — Type

SlackEvaluator{Evaluator<:AbstractNLPEvaluator, T, VT} <: AbstractNLPEvaluator

Reformulate a problem with inequality constraints as an equality constrained problem, by introducing a set of slack variables.

Description

A SlackEvaluator takes as input an original AbstractNLPEvaluator, subject to inequality constraints

\[\begin{aligned} \min_{u \in \mathbb{R}^n} \quad & f(u)\\ \mathrm{s.t.} \quad & h^♭ ≤ h(u) ≤ h^♯,\\ & u^♭ ≤ u ≤ u^♯. \end{aligned}\]

The SlackEvaluator instance rewrites this problem with inequalities as a new problem comprising only equality constraints, by introducing $m$ slack variables $s_1, ⋯, s_m$. The new problem writes out

\[\begin{aligned} \min_{u \in \mathbb{R}^n, s \in \mathbb{R}^m} \quad & f(u)\\ \mathrm{s.t.} \quad & h(u) - s = 0 \\ & u^♭ ≤ u ≤ u^♯, \\ & h^♭ ≤ s ≤ h^♯. \end{aligned}\]

Attributes

inner::Evaluator: original evaluator

s_min::VT: stores lower bounds for slack variables

s_max::VT: stores upper bounds for slack variables

nv::Int: number of original variables

ns::Int: number of slack variables

AugLagEvaluator

Argos.AugLagEvaluator — Type

AugLagEvaluator{Evaluator<:AbstractNLPEvaluator, T, VT} <: AbstractPenaltyEvaluator

Augmented-Lagrangian evaluator.

Description

Takes as input any AbstractNLPEvaluator encoding a non-linear problem

\[\begin{aligned} \min_u \quad & f(u)\\ \mathrm{s.t.} \quad & h^♭ ≤ h(u) ≤ h^♯,\\ & u^♭ ≤ u ≤ u^♯, \end{aligned}\]

and return a new evaluator reformulating the original problem by moving the $m$ constraints $h^♭ ≤ h(u) ≤ h^♯$ into the objective using a set of penalties $ϕ_1, ⋯, ϕ_m$ and multiplier estimates $λ_1, ⋯, λ_m$:

\[\begin{aligned} \min_u \quad & f(u) + \sum_{i=1}^m ϕ_i(h_i, λ_i) \\ \mathrm{s.t.} \quad & u^♭ ≤ u ≤ u^♯, \end{aligned}\]

This evaluator considers explicitly the inequality constraints, without reformulating them by introducing slack variables. Each penalty $ϕ_i$ is defined as

\[ϕ_i(h_i, λ_i) = λ_i^⊤ φ_i(h_i) + \frac \rho2 \| φ_i(h_i) \|_2^2\]

with $φ_i$ a function to compute the current infeasibility

\[φ_i(h_i, λ_i) = \max\{0 , λ_i + ρ (h_i - h_i^♯) \} + \min\{0 , λ_i + ρ (h_i - h_i^♭) \}\]

Attributes

inner::Evaluator: original problem.

cons_type: type of the constraints of the original problem (equalities or inequalities).

cons::VT: a buffer storing the current evaluation of the constraints for the inner evaluator.

rho::T: current penalty.

λ::VT: current multiplier.

scaler::MaxScaler{T,VT}: a scaler to rescale the range of the constraints in the original problem.

MOIEvaluator

The bridge to MathOptInterface is encoded by the MOIEvaluator structure:

Argos.MOIEvaluator — Type

MOIEvaluator <: MOI.AbstractNLPEvaluator

Bridge from a AbstractNLPEvaluator to a MOI.AbstractNLPEvaluator.

Attributes

nlp::AbstractNLPEvaluator: the underlying ExaPF problem.

hash_x::UInt: hash of the last evaluated variable x

has_hess::Bool (default: false): if true, pass a Hessian structure to MOI.

ProxALEvaluator

Missing docstring.

Missing docstring for ProxALEvaluator. Check Documenter's build log for details.